backend/requirements.txt
crewai
crewai_tools
langchain_openai
python-decouple
fastapi>=0.95.0
uvicorn>=0.21.0
pydantic>=1.10.7
dspy-ai
